Move window logic to window_layout.rs
